odroid: Add X support
This patch does not add a new board to the build system, but simply
changes the settings of the X2/U2/U3, and thus disables support for
these boards.

The ODROID-X has only 1 GiB of memory, unlike the X2/U2/U3 which has 2.
Memory size is set with macros at compile-time, which means that two
different builds are necessary to boot the X and the X2/U2/U3 boards.

A new `board_type' has been added to `get_board_type()', but the
detection logic can not distinguish between the X and X2 yet.

This patch is based on Kevin Mihelich's <kevin@archlinuxarm.org> work
done on Arch Linux ARM. [1]

[1]: https://raw.githubusercontent.com/archlinuxarm/PKGBUILDs/586e583ce2a0fd6cd426f3706bb92567bd805b51/alarm/uboot-odroid/0002-odroid-x-support.patch